+++ src/models/FilingReturnModelBasic.cs
@@ -67,7 +67,7 @@ namespace Avalara.AvaTax.RestClient
         /// The filing period represents the year and month of the last day of taxes being reported on this filing. 
         /// For example, an annual tax filing for Jan-Dec 2015 would have a filing period of Dec 2015.
         /// </summary>
-        public Byte? endPeriodMonth { get; set; }
+        public Int32? endPeriodMonth { get; set; }
 
         /// <summary>
         /// The year of the filing period for this tax filing.
